Closes #121: Added invokeDynamic support for binary operations where we don't know the operand types at compile time

+++ src/jmh/resources/io/jactl/benchmarks/MonteCarloPIStatic.groovy
@@ -19,7 +19,7 @@ class MonteCarloPI {
     }
 
     int nextInt(int bound) {
-      return ((nextLong() % bound) + bound) % bound
+      return (int)(((nextLong() % bound) + bound) % bound)
     }
     
     static long rol(long i, int amt) { return (i << amt) | (i >>> -amt) }
